Split out hidraw descriptor parsing into a helper `fu_hidraw_device_parse_descriptor`
