+++ libfwupdplugin/fu-hidraw-device.h
@@ -6,6 +6,7 @@
 
 #pragma once
 
+#include "fu-hid-descriptor.h"
 #include "fu-udev-device.h"
 
 #define FU_TYPE_HIDRAW_DEVICE (fu_hidraw_device_get_type())
@@ -27,3 +28,6 @@ fu_hidraw_device_get_feature(FuHidrawDevice *self,
 			     gsize bufsz,
 			     FuIoctlFlags flags,
 			     GError **error) G_GNUC_WARN_UNUSED_RESULT G_GNUC_NON_NULL(1, 2);
+FuHidDescriptor *
+fu_hidraw_device_parse_descriptor(FuHidrawDevice *self, GError **error) G_GNUC_WARN_UNUSED_RESULT
+    G_GNUC_NON_NULL(1);
